move parser code out from main
